requirements.txt
discord.py==1.7.3
Flask==2.0.1
dnspython==2.1.0
pymongo==3.12.0
pytz==2021.1
requests==2.25.1
youtube-dl==2021.6.6